requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv
gradio
